PyLEDSimulator/requirements.txt
pygame>=2.0.0
pillow>=8.0.0
numpy>=1.19.0